requirements.txt
Faker
Flask
Flask-WTF
pony
python-slugify
requests
